requirements.txt
streamlit>=1.28.0
tensorflow>=2.10.0
pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
plotly>=5.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
opencv-python>=4.5.0
scikit-learn>=1.0.0
scikit-image>=0.19.0
flatbuffers==23.5.26
fonttools==4.25.0
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.0
h5py>=3.10.0
idna==3.4
keras>=2.10.0
Keras-Preprocessing==1.1.2
libclang==16.0.6
Markdown==3.5
MarkupSafe==2.1.3
mkl-service==2.4.0
munkres==1.1.4
oauthlib==3.2.2
opt-einsum==3.3.0
ply==3.11
protobuf>=3.19.6
pyasn1-modules==0.3.0
PyQt5==5.15.7
pywin32==305.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator>=2.10.0
tensorflow-io-gcs-filesystem>=0.31.0
termcolor==2.3.0
optree>=0.17.0
wrapt>=1.17.0
